Pull crypto into its own file
